Add files via upload
Temperature flow - Visualization Project
The required task is to visualize the temperature change inside a room. The application allows
the user to build the room structure by clicking on the room area (panel/simpleOpenGlControl)
such that each click change the clicked cell into one of the following cell types:
- Normal Cell
This cell represents default cell room with temperature of Tr . The temperature of this
kind of cells is modified, over time, by averaging the temperatures of the rounding cells
together with its own temperature.
- Air Conditioner Cell
This cell represents an air conditioner that has a permanent temperature of Tc . The
temperature of this kind of cells does not change over time.
- Heater Cell
This cell represents a heater that has a permanent temperature of Th . The temperature of
this kind of cells does not change over time.
- Block Cell
This cell represents a wall that is heat-proof. The temperature of this kind of cells neither
changes over time no